junit refactor, move HSQLDB init methods into DbUtil class
